Added double type amount instead of int

+++ lib/models/payment/add_money_wallet_request.dart
@@ -1,7 +1,7 @@
 class AddMoneyToWalletRequest {
   int? customerProfileId;
   int? paymentProfile;
-  int? amount;
+  double? amount;
   String? paymentMethod;
 
   AddMoneyToWalletRequest(
